# docs/contratos/api/admin-v1.openapi.yaml
openapi: 3.1.0
info:
  title: ModelCairn Administration API
  version: 1.0.0-alpha.1
  description: |
    Draft Milestone 3 contract. Administrative responses use Cache-Control: no-store.
    Login and mutations require the configured exact Origin; online CLI clients
    supply it too. Authenticated operations require session and CSRF except
    session/me, which requires the same-origin recovery checks in admin-runtime-v1.
    Authentication failures return 401; Origin/CSRF failures return 403 before
    business mutations. These boundary errors apply to all protected operations.
servers:
  - url: /api/v1/admin
security:
  - adminSession: []
    csrfToken: []
paths:
...
  /requests:
    get:
      operationId: listOperationalRequests
      description: Content-free operational history ordered by start time and immutable request id.
      parameters:
        - { name: outcome, in: query, schema: { type: string, enum: [success, error, partial, cancelled, indeterminate] } }
        - { name: alias, in: query, schema: { type: string, maxLength: 128 } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: cursor, in: query, schema: { type: string, maxLength: 512 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200': { description: Bounded page, content: { application/json: { schema: { $ref: '#/components/schemas/OperationalRequestPage' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /requests/{id}/attempts:
    get:
      operationId: listOperationalAttempts
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, maxLength: 128 } }
      responses:
        '200':
          description: At most 32 attempts ordered by sequence.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties: { items: { type: array, maxItems: 32, items: { $ref: '#/components/schemas/OperationalAttempt' } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    adminSession: { type: apiKey, in: cookie, name: mc_session }
    csrfToken: { type: apiKey, in: header, name: X-CSRF-Token }
...
  responses:
...
    Error:
      description: Structured error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
...
    OperationalRequestPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/OperationalRequest' } }
        nextCursor: { type: [string, 'null'], maxLength: 512 }
    OperationalRequest:
      type: object
      additionalProperties: false
      required: [id, requestedAlias, startedAt, attempts]
      properties:
        id: { type: string, maxLength: 128 }
        requestedAlias: { type: string, maxLength: 128 }
        startedAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        outcome: { type: [string, 'null'], enum: [success, error, partial, cancelled, indeterminate, null] }
        httpStatus: { type: [integer, 'null'], minimum: 100, maximum: 599 }
        inputTokens: { type: [integer, 'null'], minimum: 0 }
        outputTokens: { type: [integer, 'null'], minimum: 0 }
        ttftMillis: { type: [integer, 'null'], minimum: 0 }
        durationMillis: { type: [integer, 'null'], minimum: 0 }
        attempts: { type: integer, minimum: 0, maximum: 32 }
    OperationalAttempt:
      type: object
      additionalProperties: false
      required: [sequence, startedAt, outcome, retryable]
      properties:
        sequence: { type: integer, minimum: 1, maximum: 32 }
        destinationID: { type: [string, 'null'] }
        startedAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        outcome: { type: string, enum: [success, error, partial, cancelled, indeterminate] }
        errorClass: { type: [string, 'null'] }
        providerStatus: { type: [integer, 'null'], minimum: 100, maximum: 599 }
        providerRequestID: { type: [string, 'null'] }
        retryable: { type: boolean }
        fallbackReason: { type: [string, 'null'] }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            field: { type: [string, 'null'] }
            retryable: { type: boolean, default: false }